fix(orchestration): resolve import cycle by moving AgentPerformance to orchestration.types (#6831)
- Move AgentPerformance dataclass from enhanced_orchestration/types.py to orchestration/types.py
- Update orchestration/performance_tracker.py to import AgentPerformance from orchestration.types
- Re-export AgentPerformance from enhanced_orchestration/types.py for backwards compatibility
- Eliminates circular dependency: orchestration.performance_tracker -> enhanced_orchestration.types -> orchestration.types

Fixes #6831
